finish server movie

+++ src/main/java/com/movie/backend/service/BookingService.java
@@ -5,6 +5,8 @@ import com.movie.backend.dto.BookingDTO;
 import com.movie.backend.dto.ComboDTO;
 import com.movie.backend.entity.*;
 import com.movie.backend.exception.BookingException;
+import com.movie.backend.exception.EventValidException;
+import com.movie.backend.exception.UserException;
 import com.movie.backend.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
@@ -54,7 +56,7 @@ public class BookingService {
         String seats = bookingDTO.getSeats() ;
         String[] seatNames = seats.split(", ");
 
-        // checking booking with all the name of seat from client which sent the request
+        // check booking with all the name of seat from client which sent the request
         for (String seatName : seatNames) {
             if(seatService.checkPaidBooking(eventId , seatName)) {
                 throw new BookingException("Đã có người thanh toan , vui long đặt chỗ khác");
@@ -65,9 +67,13 @@ public class BookingService {
             }
             // handle quick booking than current user
         }
-        // find all  the class was reference with booking include: user , event , ...
-        Event event = eventRepository.findById(eventId).orElseThrow();
-        User user = userRepository.findById(bookingDTO.getUser().getId()).orElseThrow();
+        // find all the entity was reference with booking include: user , event
+
+        Event event = eventRepository.findById(eventId).orElseThrow(() -> new EventValidException("Id of event not found"));
+
+        User user = userRepository.findById(bookingDTO.getUser().getId()).orElseThrow(() -> new UserException("Id of user not found"));
+
+        // save booking
         Booking booking = Booking.builder()
                 .user(user)
                 .seats(seats)
@@ -77,8 +83,6 @@ public class BookingService {
                 .build();
 
         // handle  seat with  list seat from bookingDTO request
-
-
         List<Seat> seatList = seatRepository.findByNameRoom(seatNames,
                 event.getRoom().getId() ) ;
         List<BookingSeat> bookingSeats = new ArrayList<>() ;
@@ -93,7 +97,6 @@ public class BookingService {
         });
 
         // handle all combo with  list combo of bookingDTO from request
-
         List<ComboDTO> comboOnBooking = bookingDTO.getCombos();
 
         List<BookingCombo> bookingCombos = new ArrayList<>() ;
@@ -105,7 +108,8 @@ public class BookingService {
                     .quantity(comboDTO.getQuantity())
                     .build();
             bookingCombos.add(bookingCombo);
-        } );
+        });
+
         // save to db
         Booking savedBooking = bookingRepository.save(booking);
         bookingSeatRepository.saveAll(bookingSeats);
@@ -114,6 +118,8 @@ public class BookingService {
         return savedBooking.getId();
     }
 
+    // Clean up all booking request which was not click cancel booking (out of page)
+    // This schedule run every 0.1s
     @Scheduled(fixedDelay = 100)
     public void scheduleFixedBookingRequest() {
         List<Booking> bookings = bookingRepository.findAllBookingByDay();
@@ -130,6 +136,9 @@ public class BookingService {
     }
 
     public void deleteByBookingId(Long bookingId) {
+        if(bookingId == null) {
+            throw new BookingException("Cant get the id of booking");
+        }
         bookingSeatRepository.deleteByBooking(bookingId);
         bookingComboRepository.deleteByBooking(bookingId);
         bookingRepository.deleteBookingById(bookingId);
